+++ webpack.config.js
@@ -26,7 +26,7 @@ module.exports = {
         use: {
           loader: 'babel-loader',
         },
-        exclude: path.resolve(__dirname, 'node_module'),
+        exclude: path.resolve(__dirname, 'node_modules'),
         include: path.resolve(__dirname, 'src')
       },
     ]